* develop:
  Revert "使WKWebView可以自由扩大缩小"
  update
  update README.md
  修改工程设置
  [fix]“我”的following/followers 没有读取数据库的问题
  使WKWebView可以自由扩大缩小
  [fix] 刷新问题
  update
  [fix]修改starred 只显示一条记录的问题
  消除一个警告
  二维码背景模糊截图，模拟器上出现截图动画效果，真机上没有，现在改成一致，都没有效果
  消除一个警告
  subscribe news 取消行数的限制，避免时间显示不出来的问题

# Conflicts:
#	Scarecrow.xcodeproj/project.pbxproj

+++ Scarecrow/Repos/View/ADReposTableViewCell.m
@@ -71,7 +71,7 @@
     self.nameLabel.attributedText = viewModel.name;
     self.decriptionLabel.attributedText = viewModel.reposDescription;
     self.languageLabel.text = viewModel.language;
-    self.forkCountLabel.text = [NSString stringWithFormat:@"%ld", viewModel.repos.forksCount];
+    self.forkCountLabel.text = [NSString stringWithFormat:@"%lu", (unsigned long)viewModel.repos.forksCount];
     self.updateTimeLabel.text = viewModel.udpateTime;
 }
